+++ src/game_engine/constants.py
@@ -1,8 +1,0 @@
-from enum import IntEnum
-
-
-class Difficulty(IntEnum):
-    EASY = 0
-    MEDIUM = 1
-    HARD = 2
-    LUDICROUS = 3